Refactor: alterado responsabilidade dos services e criado hooks para content
