+++ src/components/ContentForm/ContentForm.tsx
@@ -16,14 +16,14 @@ import {
   CommandItem,
   CommandList,
 } from "../ui/command";
-import { useContentTags } from "@/service/content/contentTag/getContentTag";
-import { useContentTypes } from "@/service/content/contentType/getContentType";
 import { extractBase64Images } from "../../utils/extractBase64Images";
 import { Input } from "../ui/input";
 import { useRouter } from "@tanstack/react-router";
 import type { ContentPayload } from "@/Interface/content/ContentPayload";
 import { useCreateContent } from "@/hooks/content/useCreateContent";
 import { useUpdateContent } from "@/hooks/content/useUpdateContent";
+import { useContentTypes } from "@/hooks/content/contentTypes/useGetContentTypes";
+import { useContentTags } from "@/hooks/content/contentTags/useGetContentTags";
 
 interface ContentFormProps {
   initialData?: getContentSchema | ContentPayload;
@@ -61,6 +61,7 @@ export default function ContentForm({
     isLoading: loadingTagData,
     isError: errorTagData,
   } = useContentTags();
+
   const {
     data: contenTypeData,
     isLoading: loadingTypeData,
